Add Gemfile.lock to gitignore
For gems (libraries), dependency constraints belong in the gemspec.
Avoid locking dependencies to only a single version and instead allow
installation of any dependency within the supplied semantic range. This
helps ensure that the semantic range is indeed compatible with the gem.

When applications consume the gem, they will use their own Gemfile.lock
and lock both this gem and its dependencies.