Add RSpec and a basic test
https://relishapp.com/rspec

Add a very basic, minimal Rails app in the spec directory. The spec runs
the CLI on the Rails app and verifies the dumped output matches the
expected result.

The dumps are compared using "snapshots" inspired by Jest snapshot
testing. This ensures each run has byte-for-byte identical results.

https://jestjs.io/docs/snapshot-testing

With the ground work implemented, new tests can be added to fill in
coverage and new features can be accompanied by tests.