add conditional logic for 204 status (#36)
Allow multiple expected status codes

To support dump methods with multiple responses that may have different
HTTP status codes, the dump argument "status_code" is now
"status_codes". It accepts an Array instead of a scalar.

To support the HTTP 204 status code (No Content), the Content-Type
header is only parsed if it exists. If it does not exist, the dumped
response does not have a file extension.