Include an HTTP status line and response headers in dumps
The dumped responses now contain an HTTP status line and HTTP headers.
Each header line ends with a carriage return and line feed. The headers
are separated from the body by an empty line consisting of a carriage
return and line feed.

This makes the dumps provide more information about the response that
can be used in client side testing. For example, the status code or
content type might be useful when interpreting the dump file.