Fix --fail-fast option
The first implementation was only breaking early out of the loop that
iterated through a single dumper's dumps.

However, it would then move on to the next dumper as usual.

This update will exit both loops and correctly abort after first
failure.

Added a second dumper to the fail_app specs to verify this was
working correctly.