Support multipart/form-data submission stubbing (#84)
For the first time we are now attaching files to form submissions /
fetches /
POSTs. In order to test this, we need to be able to mock and stub such
fetches.
File-containing submissions take the form of a "multipart/form-data"
request rather than
an "urlencoded" request. More details can be found here:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods/POST. When a
multipart
request contains a binary file in its body (such as a PNG), this can
cause
encoding difficulties. In our case, runner.rb uses
JSON.pretty_generate() on
the entire result of every dump, which includes that binary-containing
body,
causing an error when .pretty_generate() cannot convert it to UTF-8
encoding as
it requires.

The chosen solution has changed the dump output structure such that
request and response
bodies are both always given their own separate file, .request_body and
.response_body.
These are written with in a way that is safe to run on possible
binaries, but also
should keep the more common, nonbinary bodies easily readable in a text
editor.

---------

Co-authored-by: Gavin Myers <gavinl@pvep.com>