Change Dependabot update schedule to weekly
Dependabot was previously configured to run daily, which created a large amount
of PR noise for routine dependency updates, when most of these updates are not
time-critical. Further, a reduction in CI runs on Dependabot updates could
result in fewer GitHub minutes used, potentially reducing some cost over the
year.

Refs https://github.com/Pioneer-Valley-Books/literacyfootprints/issues/14862